// src/modules/admin/admin.interface.ts
export interface IAdminUserFilter {
    page?: number;
    limit?: number;
    role?: string;
    status?: string;
}

export interface IAdminPropertyFilter {
    page?: number;
    limit?: number;
    userId?: string;
}

export interface IAdminRentalFilter {
    page?: number;
    limit?: number;
    status?: string;
}